O.K. we have corporates.
I've decided to ignore roads when creating a corporate, instead, I just chopped the whole board to equal squares and each building was assigned to a corporation according to the location of it's center.
after that, I've called 10 random corporates and each could try and take-over his adjacent corporates (success rate of a takeover is about 20% per takeover, which gives us somewhere between 10 to 20 successful takeovers.